apis create, update, delete, get dimention

+++ src/main/java/com/higgsup/base/service/impl/UserService.java
@@ -87,17 +87,25 @@ public class UserService implements IUserService {
 
     @Override
     public List<DimensionDTO> getDimensions(Long userId, Integer dimensionNumber) {
-        List<DimensionDTO> dimensionDTOS = new ArrayList<>();
-        List<Dimention> dimentions = dimentionRepository.getDimentions(userId, dimensionNumber);
-
-        // convert from dimension entity to dimention DTO by bean copy
-        for (Dimention dimention : dimentions) {
-            DimensionDTO dimensionDTO = new DimensionDTO();
-            BeanUtils.copyProperties(dimention, dimensionDTO);
-            dimensionDTOS.add(dimensionDTO);
-
+        List<DimensionDTO> dimensionDTOS =new ArrayList<>();
+        List<Dimention> dimentions;
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (!userOptional.isPresent()){
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND, String.valueOf(ErrorCode.USER_NOT_FOUND.getErrorCode()));
+        } else {
+            if (dimensionNumber == null){
+                dimentions = dimentionRepository.findAllByUserId(userId);
+            } else {
+                dimentions = dimentionRepository.getDimentions(userId, dimensionNumber);
+            }
+            // convert from dimension entity to dimention DTO by bean copy
+            for (Dimention dimention : dimentions) {
+                DimensionDTO dimensionDTO = new DimensionDTO();
+                BeanUtils.copyProperties(dimention, dimensionDTO);
+                dimensionDTOS.add(dimensionDTO);
+            }
+            return dimensionDTOS;
         }
-        return dimensionDTOS;
     }
 
     @Override
@@ -200,33 +208,54 @@ public class UserService implements IUserService {
                 }
                 user.setPassword(passwordEncoder.encode(newPassword));
                 return userRepository.save(user) != null;
-
             }
         }
     }
 
     @Override
     public DimensionDTO saveDimension(DimensionDTO dimensionDTO, Long userId) {
-        Dimention dimention = mapperFacade.map(dimensionDTO, Dimention.class);
-        dimention.setLastUpdated(new Timestamp(System.currentTimeMillis()));
-        dimention.setUserId(userId);
-        return mapperFacade.map(dimentionRepository.save(dimention), DimensionDTO.class);
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (!userOptional.isPresent()){
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND, String.valueOf(ErrorCode.USER_NOT_FOUND.getErrorCode()));
+        }else {
+            Dimention dimention = mapperFacade.map(dimensionDTO, Dimention.class);
+            dimention.setLastUpdated(new Timestamp(System.currentTimeMillis()));
+            dimention.setUserId(userId);
+            return mapperFacade.map(dimentionRepository.save(dimention), DimensionDTO.class);
+        }
     }
 
     @Override
-    public DimensionDTO updateDimension(DimensionDTO dimensionDTO) {
-        Dimention dimention = mapperFacade.map(dimensionDTO, Dimention.class);
-        dimention.setLastUpdated(new Timestamp(System.currentTimeMillis()));
-        return mapperFacade.map(dimentionRepository.save(dimention), DimensionDTO.class);
+    public DimensionDTO updateDimension(Long userId, DimensionDTO dimensionDTO, Long dimensionId) {
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (!userOptional.isPresent()){
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND, String.valueOf(ErrorCode.USER_NOT_FOUND.getErrorCode()));
+        }else {
+            Optional<Dimention> dimensionOptional = dimentionRepository.findById(dimensionId);
+            if (!dimensionOptional.isPresent()){
+                throw new BusinessException(ErrorCode.DIMENSION_IS_EMPTY, String.valueOf(ErrorCode.DIMENSION_IS_EMPTY.getErrorCode()));
+            }else {
+                Dimention dimension = dimensionOptional.get();
+                dimension.setLastUpdated(new Timestamp(System.currentTimeMillis()));
+                BeanUtils.copyProperties(dimensionDTO, dimension, "id");
+                dimentionRepository.save(dimension);
+                return dimensionDTO;
+            }
+        }
     }
 
     @Override
-    public void deleteDimension(Long dimensionId) {
-        Optional<Dimention> dimentionOptional = dimentionRepository.findById(dimensionId);
-        if(!dimentionOptional.isPresent()) {
-            throw new RuntimeException(String.valueOf(ErrorCode.VALIDATION.getErrorCode()));
-        }else{
-            dimentionRepository.delete(dimentionOptional.get());
+    public void deleteDimension(Long dimensionId, Long userId) {
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (userOptional.isPresent()){
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND, String.valueOf(ErrorCode.USER_NOT_FOUND.getErrorCode()));
+        } else {
+            Optional<Dimention> dimentionOptional = dimentionRepository.findById(dimensionId);
+            if(!dimentionOptional.isPresent()) {
+                throw new BusinessException(ErrorCode.DIMENSION_IS_EMPTY, String.valueOf(ErrorCode.DIMENSION_IS_EMPTY.getErrorCode()));
+            }else{
+                dimentionRepository.delete(dimentionOptional.get());
+            }
         }
     }
 }
